+++ Nav.php
@@ -176,7 +176,7 @@ class Nav extends Widget
             $active = $this->isItemActive($item);
         }
 
-        if ($items !== null) {
+        if (!empty($items)) {
             $linkOptions['data-toggle'] = 'dropdown';
             Html::addCssClass($options, ['widget' => 'dropdown']);
             Html::addCssClass($linkOptions, ['widget' => 'dropdown-toggle']);
@@ -189,6 +189,10 @@ class Nav extends Widget
                 }
                 $items = $this->renderDropdown($items, $item);
             }
+        }
+
+        if (empty($items)) {
+            $items = '';
         }
 
         if ($this->activateItems && $active) {